Split parsed string classes into new module
